re #2 Allow alternate AllowedCallersFilter to be used in JWTSecurityConfigurer
