+++ pom.xml
@@ -31,7 +31,7 @@
         <version.log4j-1.2-api>2.11.2</version.log4j-1.2-api>
         <version.metrics-spring>3.1.3</version.metrics-spring>
         <version.microservice.accumulo-utils>1.3</version.microservice.accumulo-utils>
-        <version.microservice.authorization.api>1.4</version.microservice.authorization.api>
+        <version.microservice.authorization.api>1.7-SNAPSHOT</version.microservice.authorization.api>
         <version.microservice.base-rest-responses>1.2</version.microservice.base-rest-responses>
         <version.microservice.cache-starter>1.0</version.microservice.cache-starter>
         <version.microservice.metadata-utils>1.3</version.microservice.metadata-utils>
